WriteGen-backend-firebase/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
firebase-admin==6.0.1
google-cloud-firestore==2.11.1
google-cloud-storage==2.13.0
python-multipart==0.0.6
pydantic==2.5.3
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
potrace==1.0.0
svgwrite==1.4.3
reportlab==4.0.9
aiofiles==23.2.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
